config/requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.8.30
dj-database-url==2.2.0
django==5.1.1
django-cors-headers==4.4.0
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
pillow==10.4.0
psycopg2-binary==2.9.9
pyjwt==2.9.0
python-telegram-bot==21.5
redis==5.0.8
sniffio==1.3.1
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.1
